LocationService done
Built a location service that listens for the location in the
background, at an interval.
Build a HashMap to store all the markers and update only the position when location changed, if the marker already exists.
Might do a way to show only online users, or markers in a radius
from the user